requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
pydantic==2.10.6
python-multipart==0.0.20
python-dotenv==1.0.1
aiofiles==24.1.0